Reorganize file structure, add gitignore
